Update suite to v1.1 in-place, add changelog, update README and copyrights
- Replace versioned file names with canonical names; v1.1 FSCX_REVOLVE_N
  changes are now the authoritative content of the suite files
- Add CHANGELOG.md documenting v1.1, v1.0, and earlier history
- Update README with FSCX_REVOLVE_N definition, nonce derivation rules,
  and mathematical proof that HKEX equality is preserved under the nonce
- Update copyright year ranges to 2026 across all source files
