+++ CHANGELOG.md
@@ -1,0 +1,70 @@
+# Changelog
+
+All notable changes to the Herradura Cryptographic Suite are documented here.
+
+---
+
+## [1.1] - 2026
+
+### Added
+- **`FSCX_REVOLVE_N`** primitive in `Herradura cryptographic suite` (Go and Python).
+  Each iteration XORs a nonce N into the result:
+  ```
+  result = FSCX(result, B) ⊕ N
+  ```
+  This converts `FSCX_REVOLVE` from a purely GF(2)-linear function to an affine
+  function, breaking per-step linearity while preserving the HKEX equality and
+  orbit properties. See the [FSCX_REVOLVE_N section in the README](README.md)
+  for the mathematical proof.
+
+- **Session-specific nonce derivation** (no new secrets required):
+  - HKEX, HPKS, HPKE: `hkex_nonce = C ⊕ C2`, computable from the public key
+    since A2 and B2 are public (C2 = fscx_revolve(A2, B2, i)).
+  - HSKE: nonce = preshared key, injecting the key at every revolve step rather
+    than only at the input/output boundaries.
+
+### Changed
+- `Herradura cryptographic suite.go` and `Herradura cryptographic suite.py`:
+  replaced all `fscx_revolve` calls (except the initial public-value generation
+  of C and C2) with `fscx_revolve_n` using the appropriate nonce.
+- Copyright year ranges updated across all source files to include 2026.
+
+---
+
+## [1.0] - 2024
+
+### Added
+- Initial release of `Herradura cryptographic suite` (Go and Python) implementing:
+  - **HKEX** – key exchange in the style of Diffie-Hellman, using `FSCX_REVOLVE`.
+  - **HSKE** – symmetric key encryption: `E = fscx_revolve(P, key, i)`;
+    decrypt with `P = fscx_revolve(E, key, r)`.
+  - **HPKS** – public key signature (must be composed with HSKE to be secure).
+  - **HPKE** – public key encryption.
+  - EVE bypass test suite covering all four protocols.
+
+---
+
+## Earlier history
+
+### ARM assembly example – 2023
+- Added `HKEX_arm_linux.s`: HKEX in ARM 32-bit assembly for Linux (Cortex-A7,
+  thumb mode), runnable via QEMU.
+
+### Bug fixes and cleanup – 2024
+- Fixed `getMax` function in the C sample (`Herradura_KEx.c`).
+- General code cleanup across C implementations.
+
+### Python HKEX refactor – 2024
+- Refactored `Herradura_KEx.py` for clarity and correctness (argument parsing,
+  type hints, secure random generation with `secrets.randbits`).
+
+---
+
+## Original samples – 2017
+
+- `Herradura_KEx.c` – reference C implementation of HKEX (64-bit integers).
+- `Herradura_KEx_bignum.c` – arbitrary-precision HKEX using GNU MP (libgmp).
+- `Herradura_AEn.c` – HAEN asymmetric encryption (deprecated; use HSKE instead).
+- `Herradura_KEx.go` – Go HKEX sample using `math/big`.
+- `FSCX_HAEN1.ino` / `FSCX_HAEN1_ulong.ino` – Arduino proofs of concept (16/32-bit).
+- `HAEN.asm` – HAEN in Intel x86 32-bit NASM assembly.
