+++ FSCX_HAEN1_ulong.ino
@@ -1,6 +1,6 @@
 /*  HKEX and HAEN 32 bit proof of concept code for Arduino 
 
-    Copyright (C) 2021 Omar Alejandro Herrera Reyna
+    Copyright (C) 2021-2026 Omar Alejandro Herrera Reyna
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the MIT License or the GNU General Public License
