+++ Herradura cryptographic suite.py
@@ -1,11 +1,11 @@
 '''
-    Herradura Cryptographic Suite
-    
-    Copyright (C) 2024 Omar Alejandro Herrera Reyna
+    Herradura Cryptographic Suite v1.1
+
+    Copyright (C) 2024-2026 Omar Alejandro Herrera Reyna
 
     This program is free software: you can redistribute it and/or modify
-    it under the terms of the MIT License or the GNU General Public License 
-    as published by the Free Software Foundation, either version 3 of the License, 
+    it under the terms of the MIT License or the GNU General Public License
+    as published by the Free Software Foundation, either version 3 of the License,
     or (at your option) any later version.
 
     Under the terms of the GNU General Public License, please also consider that:
@@ -16,7 +16,31 @@
     GNU General Public License for more details.
 
     You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <https://www.gnu.org/licenses/>. */
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+    --- v1.1: FSCX_REVOLVE_N ---
+
+    v1.1 introduces FSCX_REVOLVE_N: a nonce-augmented variant of FSCX_REVOLVE
+    where each iteration XORs a nonce N after the FSCX step:
+        result = FSCX(result, B) ^ N
+
+    This converts the purely linear GF(2) function to affine, breaking linearity
+    while preserving the HKEX equality and orbit properties.
+
+    Nonce derivation (no new secrets):
+    - For HKEX, HPKS, HPKE: hkex_nonce = C ^ C2  — computable from the public key
+      (C is in the public key; C2 = fscx_revolve(A2, B2, i) can be computed from
+      A2, B2 also in the public key)
+    - For HSKE: N = preshared key — the key is injected at every revolve step,
+      not just at input/output boundaries
+
+    Mathematical proof that HKEX equality is preserved:
+    The HKEX equality
+        FSCX_REVOLVE_N(C2, B, N, r) ^ A = FSCX_REVOLVE_N(C, B2, N, r) ^ A2
+    holds because when you expand with C = FSCX_REVOLVE(A, B, i) and
+    C2 = FSCX_REVOLVE(A2, B2, i), and use L^(r+i) = I (since r+i=P), the
+    condition reduces to L^r(T_i(Z)) = T_r(Z) — the same condition as without
+    the nonce. N cancels identically from both sides.
 '''
 
 from bitstring import BitArray
@@ -35,6 +59,18 @@ def fscx_revolve(A, B, steps, verbose=False):
             print(f"Step {step + 1}: {result.hex}")
     return result
 
+def fscx_revolve_n(A, B, nonce, steps, verbose=False):
+    """Nonce-augmented FSCX_REVOLVE (v1.1).
+    Each step: result = fscx(result, B) ^ nonce
+    Breaks pure GF(2)-linearity while preserving HKEX equality and orbit properties.
+    """
+    result = A.copy()
+    for step in range(steps):
+        result = fscx(result, B) ^ nonce
+        if verbose:
+            print(f"Step {step + 1}: {result.hex}")
+    return result
+
 def fscx(A, B):     # [binary] full surroundings cyclic xor
     result = A ^ B
     A.rol(1)
@@ -47,7 +83,7 @@ def fscx(A, B):     # [binary] full surroundings cyclic xor
     B.rol(1)
     return result
 
-'''  
+'''
 let,    Alice, Bob: i + r == bitlength b;  i == 1/4 bitlength; r == 3/4 bitlength; bitlength is a power of 2 >= 8
         P be a plaintext message of bitlength b,
         E the encrypted version of plaintext P,
@@ -56,7 +92,7 @@ let,    Alice: A,B be random values of bitlength b,
         Bob: A2,B2 be random values of bitlength b
 let,    Alice: C = fscx_revolve(A, B, i) ,
         Bob: C2 = fscx_revolve(A2, B2, i)
-then,   Alice: D = fscx_revolve(C2, B, r) ^ A , 
+then,   Alice: D = fscx_revolve(C2, B, r) ^ A ,
         Bob: D2 = fscx_revolve(C, B2, r) ^ A2
 where,  Alice, Bob: D == D2
 then,   fscx_revolve(C2, B, r) ^ A  == fscx_revolve(C, B2, r) ^ A2,
@@ -70,7 +106,7 @@ then,   E = fscx_revolve(C, B2, r) ^ A2  ^ P,
         P == (D = fscx_revolve(C2, B, r) ^ A ^ E)
 
 let,    E = fscx_revolve(C2, B, r) ^ A  ^ P
-then,   fscx_revolve(E, B2, i) ^ A2 ^ P  == 0  
+then,   fscx_revolve(E, B2, i) ^ A2 ^ P  == 0
         fscx_revolve(E ^ P, B2, i) == 0
 
 HKEX (key exchange)
@@ -84,9 +120,9 @@ HKEX (key exchange)
 HSKE (symmetric key encryption):
     Alice,Bob:  share key of bitlength b
     Alice:  E = fscx_revolve(P , key , i)
-            shares E with Bob 
+            shares E with Bob
     Bob:    P = fscx_revolve(E , key , r)
-        
+
 HPKS (public key signature)
     Alice:  C = fscx_revolve(A,B,i)
             C2 = fscx_revolve(A2,B2,i)
@@ -101,7 +137,7 @@ HPKE (public key encryption)
             {publish (C,B2,A2,r) as public key, keep the rest of parameters as private key},
     Bob:    E = fscx_revolve(C, B2, r) ^ A2  ^ P
             shares E with Alice
-    Alice:  P = fscx_revolve(C2, B, r) ^ A ^ E             
+    Alice:  P = fscx_revolve(C2, B, r) ^ A ^ E
 '''
 # Examples with b = 256 bits:
 r_value = 192  # Adjust as needed
@@ -119,6 +155,8 @@ C = fscx_revolve(A, B, i_value)
 print(f"C         : {C.hex}")
 C2 = fscx_revolve(A2, B2, i_value)
 print(f"C2        : {C2.hex}")
+hkex_nonce = C ^ C2
+print(f"hkex_nonce: {hkex_nonce.hex}")
 nonce = new_rand_bitarray(256)
 print(f"nonce     : {nonce.hex}")
 preshared = new_rand_bitarray(256)
@@ -127,9 +165,9 @@ plaintext = new_rand_bitarray(256)
 print(f"plaintext : {plaintext.hex}")
 
 print (f"\n--- HKEX (key exchange)")
-skeyA = fscx_revolve(C2 , B, r_value) ^ A
+skeyA = fscx_revolve_n(C2 , B, hkex_nonce, r_value) ^ A
 print(f"skeyA (Alice): {skeyA.hex}")
-skeyB = fscx_revolve(C , B2, r_value) ^ A2
+skeyB = fscx_revolve_n(C , B2, hkex_nonce, r_value) ^ A2
 print(f"skeyB (Bob)  : {skeyB.hex}")
 if skeyA == skeyB: # Assert equality
     print("+ session keys skeyA and skeyB are equal!")
@@ -137,19 +175,19 @@ else:
     print("- session keys skeyA and skeyB are different!")
 
 print ("\n--- HSKE (symmetric key encryption)")
-E = fscx_revolve(plaintext,  preshared , i_value)  
+E = fscx_revolve_n(plaintext, preshared, preshared, i_value)
 print(f"E (Alice) : {E.hex}")
-D = fscx_revolve(E , preshared , r_value)
+D = fscx_revolve_n(E , preshared, preshared, r_value)
 print(f"D (Bob)   : {D.hex}")
 if D == plaintext: # Assert equality
-    print("+ plaintext is correctly decrypted from E with preshared key!")
+    print("+ plaintext is correctly decrypted from E with preshared key")
 else:
     print("- plaintext is different from decrypted E with preshared key!")
 
 print ("\n--- HPKS (public key signature)")
-S = fscx_revolve(C2 , B, r_value) ^ A  ^ plaintext
+S = fscx_revolve_n(C2 , B, hkex_nonce, r_value) ^ A  ^ plaintext
 print(f"S (Alice) : {S.hex}")
-V = fscx_revolve(C, B2, r_value) ^ A2 ^ S  
+V = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ S
 print(f"V (Bob)   : {V.hex}")
 if V == plaintext: # Assert equality
     print("+ signature S from plaintext is correct!")
@@ -157,13 +195,13 @@ else:
     print("- signature S from plaintext is incorrect!")
 
 print ("\n--- HPKS (public key signature) + HSKE (symmetric key encryption) with preshared key made public")
-E = fscx_revolve(plaintext, preshared, i_value)
+E = fscx_revolve_n(plaintext, preshared, preshared, i_value)
 print(f"E (Alice) : {E.hex}")
-S = fscx_revolve(C2 , B, r_value) ^ A  ^ E # A+B2+C is the trapdoor for deceiving EVE!!!!
+S = fscx_revolve_n(C2 , B, hkex_nonce, r_value) ^ A  ^ E # A+B2+C is the trapdoor for deceiving EVE!!!!
 print(f"S (Alice) : {S.hex}")
-V = fscx_revolve(C, B2, r_value) ^ A2 ^ S   # => E
+V = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ S   # => E
 print(f"V (Bob)   : {V.hex}")
-D = fscx_revolve(V, preshared, r_value)  # => plainText
+D = fscx_revolve_n(V, preshared, preshared, r_value)  # => plainText
 print(f"D (Bob)   : {D.hex}")
 if D == plaintext: # Assert equality
     print("+ signature S(E) from plaintext is correct!")
@@ -171,9 +209,9 @@ else:
     print("- signature S(E) from plaintext is incorrect!")
 
 print ("\n--- HPKE (public key encryption)")
-E = fscx_revolve(C, B2, r_value) ^ A2 ^ plaintext
+E = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ plaintext
 print(f"E (Bob)   : {E.hex}")
-D = fscx_revolve(C2 , B, r_value)  ^ A ^ E  # => plaintext
+D = fscx_revolve_n(C2 , B, hkex_nonce, r_value)  ^ A ^ E  # => plaintext
 print(f"D (Alice) : {D.hex}")
 if D == plaintext: # Assert equality
     print("+ plaintext is correctly decrypted from E with private key!")
@@ -182,9 +220,9 @@ else:
 
 print (f"\n\n*** EVE bypass TESTS")
 print (f"\n*** HPKS (public key signature)")
-S = fscx_revolve(C, B2, r_value) ^ nonce  # ^ bruteForceValue  ## w/o A+B+C2 Eve would be forced to do a Brute force attack to find it.
+S = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ nonce  # ^ bruteForceValue  ## w/o A+B+C2 Eve would be forced to do a Brute force attack to find it.
 print(f"S (Eve)   : {S.hex}")
-V = fscx_revolve(C, B2, r_value) ^ A2  # X
+V = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2  # X
 print(f"V (Bob)   : {V.hex}")
 if V == nonce: # Assert equality
     print("+ nonce fake signature 1 verification with Alice public key is correct!")
@@ -192,7 +230,7 @@ else:
     print("- nonce fake signature 1 verification with Alice public key is incorrect!")
 S2 = V ^ nonce
 print(f"S2 (Eve)  : {S2.hex}")
-V2 = fscx_revolve(C, B2, r_value) ^ A2 ^ S2 # KK
+V2 = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ S2 # KK
 print(f"V2 (Bob)  : {V2.hex}")
 if V2 == nonce: # Assert equality
     print("+ nonce fake signature 2 verification with Alice public key is correct!")
@@ -200,17 +238,17 @@ else:
     print("- nonce fake signature 2 verification with Alice public key is incorrect!")
 
 print (f"\n*** HPKS (public key signature) + HSKE (symmetric key encryption) with preshared key made public")
-E = fscx_revolve(nonce, preshared, i_value)
+E = fscx_revolve_n(nonce, preshared, preshared, i_value)
 print(f"E (Eve)   : {E.hex}")
-S = fscx_revolve(C, B2, r_value) ^ A2 ^ E  # ^ bruteForceValue  ## w/o A+B2+C  Eve would be forced to do a Brute force attack to find it.
+S = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ E  # ^ bruteForceValue  ## w/o A+B2+C  Eve would be forced to do a Brute force attack to find it.
 print(f"S (Eve)   : {S.hex}")
-V = fscx_revolve(C, B2, r_value) ^ A2 #X
+V = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 #X
 print(f"V (Eve)   : {V.hex}")
 S2 = V ^ S
 print(f"S2 (Eve)  : {S2.hex}")
-V2 = fscx_revolve(C, B2, r_value) ^ A2 ^ S2 # KK
+V2 = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ S2 # KK
 print(f"V2 (Bob)  : {V2.hex}")
-D = fscx_revolve(V2, preshared, r_value)
+D = fscx_revolve_n(V2, preshared, preshared, r_value)
 print(f"D (Bob)   : {D.hex}") #X
 if D == nonce: # Assert equality
     print("+ fake signature(encrypted nonce) verification with Alice public key is correct!")
@@ -218,17 +256,17 @@ else:
     print("- fake signature(encrypted nonce) verification with Alice public key is incorrect!")
 
 print (f"\n*** HPKS (public key signature) + HSKE (symmetric key encryption) with preshared key made public - v2")
-S = fscx_revolve(C, B2, r_value) ^ A2 ^ nonce  # ^ bruteForceValue  ## w/o A+B2+C  Eve would be forced to do a Brute force attack to find it.
+S = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ nonce  # ^ bruteForceValue  ## w/o A+B2+C  Eve would be forced to do a Brute force attack to find it.
 print(f"S (Eve)   : {S.hex}")
-E = fscx_revolve(S, preshared, i_value)
+E = fscx_revolve_n(S, preshared, preshared, i_value)
 print(f"E (Eve)   : {E.hex}")
-V = fscx_revolve(C, B2, r_value) ^ A2 #X
+V = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 #X
 print(f"V (Eve)   : {V.hex}")
 S2 = V ^ E
 print(f"S2 (Eve)  : {S2.hex}")
-V2 = fscx_revolve(C, B2, r_value) ^ A2 ^ S2 # KK
+V2 = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ S2 # KK
 print(f"V2 (Bob)  : {V2.hex}")
-D = fscx_revolve(V2, preshared, r_value)
+D = fscx_revolve_n(V2, preshared, preshared, r_value)
 print(f"D (Bob)   : {D.hex}") #X
 if D == nonce: # Assert equality
     print("+ fake signature(encrypted nonce) v2 verification with Alice public key is correct!")
@@ -236,12 +274,12 @@ else:
     print("- fake signature(encrypted nonce) v2 verification with Alice public key is incorrect!")
 
 print (f"\n*** HPKE (public key encryption)")
-E = fscx_revolve(C, B2, r_value) ^ A2 ^ plaintext  # ^ bruteForceValue  ## w/o A+B2+C  Eve would be forced to do a Brute force attack to find it.
+E = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 ^ plaintext  # ^ bruteForceValue  ## w/o A+B2+C  Eve would be forced to do a Brute force attack to find it.
 print(f"E (Bob)   : {E.hex}")
-D = fscx_revolve(C, B2, r_value) ^ A2 #X, but == fsession from private/public key generation if components had been reused from an HKEX!?
+D = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ A2 #X, but == fsession from private/public key generation if components had been reused from an HKEX!?
 print(f"D (Eve)   : {D.hex}")
 E2 = D ^ E
-D2 = fscx_revolve(C, B2, r_value) ^ E2 # KK
+D2 = fscx_revolve_n(C, B2, hkex_nonce, r_value) ^ E2 # KK
 print(f"D2 (Eve)  : {D2.hex}")
 if (D == nonce) or (D2 == nonce): # Assert equality
     print("+ Eve could decrypt plaintext without Alice's private key!")
