+++ Herradura_AEn.c
@@ -1,7 +1,7 @@
 /*  Herradura AEn (HAEN)- an asymmetric, one to one, cipher based on the FSCX function and 
     parameters from a previous Key Exchange with Herradura KEx (HKEX).
     
-    Copyright (C) 2017-2019 Omar Alejandro Herrera Reyna
+    Copyright (C) 2017-2026 Omar Alejandro Herrera Reyna
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the MIT License or the GNU General Public License
